requirements.txt
aiohttp>=3.10.6
python-socketio[asyncio_client]>=5.11.4
python-dotenv>=1.0.1
neo4j>=5.24.0
mistralai>=1.1.0
